Changed the datatype of functions from int to float

+++ FunnyBrain/kernel.cu
@@ -8,7 +8,7 @@
 const int SQRT_MAX_THREADS = static_cast<int>(sqrt(MAX_THREADS));
 const dim3 THREADS_PER_BLOCK(SQRT_MAX_THREADS, SQRT_MAX_THREADS);
 
-__global__ void CUDAAddArrays(const int* dev_a, const int* dev_b, int* dev_c, const int arrayLength)
+__global__ void CUDAAddArrays(const float* dev_a, const float* dev_b, float* dev_c, const int arrayLength)
 {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -17,7 +17,7 @@ __global__ void CUDAAddArrays(const int* dev_a, const int* dev_b, int* dev_c, co
     }
 }
 
-__global__ void CUDASubtractArrays(const int* dev_a, const int* dev_b, int* dev_c, const int arrayLength)
+__global__ void CUDASubtractArrays(const float* dev_a, const float* dev_b, float* dev_c, const int arrayLength)
 {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -26,14 +26,14 @@ __global__ void CUDASubtractArrays(const int* dev_a, const int* dev_b, int* dev_
     }
 }
 
-__global__ void CUDAMultiplyArrays(const int* dev_a, const int* dev_b, int* dev_c, 
+__global__ void CUDAMultiplyArrays(const float* dev_a, const float* dev_b, float* dev_c, 
     const int a_rows, const int a_columns, const int b_columns) {
     int row = blockDim.y * blockIdx.y + threadIdx.y;
     int column = blockDim.x * blockIdx.x + threadIdx.x;
     
     if (row < a_rows && column < b_columns) {
         // temporary variable used to compute the result
-        int h = 0;
+        float h = 0;
         for (int i = 0; i < a_columns; ++i) {
             h += dev_a[row * a_columns + i] * dev_b[i * b_columns + column];
         }
@@ -42,21 +42,21 @@ __global__ void CUDAMultiplyArrays(const int* dev_a, const int* dev_b, int* dev_
     }   
 }
 
-__host__ void AddArrays(const int* dev_a, const int* dev_b, int* dev_c, const int arrayLength) {
+__host__ void AddArrays(const float* dev_a, const float* dev_b, float* dev_c, const int arrayLength) {
     dim3 numBlocks((arrayLength / MAX_THREADS) + (arrayLength % MAX_THREADS == 0 ? 0 : 1));
-    CUDAAddArrays << < numBlocks, MAX_THREADS >> > (dev_a, dev_b, dev_c, arrayLength);
+    CUDAAddArrays <<< numBlocks, MAX_THREADS >>> (dev_a, dev_b, dev_c, arrayLength);
 }
 
-__host__ void SubtractArrays(const int* dev_a, const int* dev_b, int* dev_c, const int arrayLength) {
+__host__ void SubtractArrays(const float* dev_a, const float* dev_b, float* dev_c, const int arrayLength) {
     dim3 numBlocks((arrayLength / MAX_THREADS) + (arrayLength % MAX_THREADS == 0 ? 0 : 1));
-    CUDASubtractArrays << < numBlocks, MAX_THREADS >> > (dev_a, dev_b, dev_c, arrayLength);
+    CUDASubtractArrays <<< numBlocks, MAX_THREADS >>> (dev_a, dev_b, dev_c, arrayLength);
 }
 
-__host__ void Multiply2d(const int* dev_a, const int* dev_b, int* dev_c, 
+__host__ void Multiply2d(const float* dev_a, const float* dev_b, float* dev_c, 
     const int a_rows, const int a_columns, const int b_columns) {
     dim3 numBlocks((b_columns / MAX_THREADS) + (b_columns % MAX_THREADS == 0 ? 0 : 1),
         (a_rows / MAX_THREADS) + (a_rows % MAX_THREADS == 0 ? 0 : 1));
-    CUDAMultiplyArrays << <  numBlocks, THREADS_PER_BLOCK >> > (dev_a, dev_b, dev_c, a_rows, a_columns, b_columns);
+    CUDAMultiplyArrays <<<  numBlocks, THREADS_PER_BLOCK >>> (dev_a, dev_b, dev_c, a_rows, a_columns, b_columns);
 }
 
 __host__ void* Create(size_t sizeInBytes) {
